refactor(polykybd): move HIL force-master override to keyboard level
The POLYKYBD_HIL is_keyboard_master_impl() override was duplicated in both
the split72 and corne42 default keymaps. Move the single implementation to
the keyboard-level polykybd.c (compiled for every variant) and the
-DPOLYKYBD_HIL flag plumbing to the keyboard-level rules.mk, so there is now
one copy instead of two.

polykybd.c is the keyboard-named source, which QMK compiles automatically;
it was also listed explicitly in SRC, so once it defined a symbol the link
failed with a duplicate definition. Removed the redundant polykybd.c entry
from SRC.

No behaviour change: verified by disassembly that default builds still call
usb_vbus_state (stock VBUS detection) and HIL builds call
eeconfig_read_handedness + usb_disconnect, for both split72 and corne42.
